Reorganize implementations layout
