+++ src/compiler/deserialization/ucd-support-inset.ts
@@ -2,8 +2,8 @@ import { esImport, esMemberAccessor, esline } from 'esgen';
 import { UcPresentationName } from '../../schema/uc-presentations.js';
 import { UcSchema } from '../../schema/uc-schema.js';
 import { UC_TOKEN_INSET_URI_PARAM } from '../../syntax/uc-token.js';
-import { UcrxCore$stubBody } from '../impl/ucrx-core.stub.js';
 import { UccConfig } from '../processor/ucc-config.js';
+import { UcrxCore$stubBody } from '../rx/impl/ucrx-core.stub.js';
 import { UcrxCore } from '../rx/ucrx-core.js';
 import { UcrxSetup } from '../rx/ucrx-setup.js';
