+++ src/compiler/rx/impl/custom-base.ucrx.class.ts
@@ -1,5 +1,5 @@
 import { EsClass } from 'esgen';
-import { UcrxSignature } from '../rx/ucrx.class.js';
+import { UcrxSignature } from '../ucrx.class.js';
 import { VoidUcrxClass } from './void.ucrx.class.js';
 
 export class CustomBaseUcrxClass extends EsClass<UcrxSignature.Args> {
