+++ src/compiler/serialization/ucs-support-map.ts
@@ -15,8 +15,8 @@ import { ucOptional } from '../../schema/uc-optional.js';
 import { UcSchema } from '../../schema/uc-schema.js';
 import { UnsupportedUcSchemaError } from '../common/unsupported-uc-schema.error.js';
 import { UC_MODULE_SERIALIZER } from '../impl/uc-modules.js';
-import { ucsCheckConstraints } from '../impl/ucs-check-constraints.js';
 import { UccConfig } from '../processor/ucc-config.js';
+import { ucsCheckConstraints } from './impl/ucs-check-constraints.js';
 import { UcsFormatterContext, UcsFormatterSignature } from './ucs-formatter.js';
 import { UcsLib } from './ucs-lib.js';
 import { UcsSetup } from './ucs-setup.js';
